move custom attributes to its own folder
