+++ FabricHealer/Attributes/CustomServiceInitializerAttribute.cs
@@ -1,15 +1,15 @@
 ﻿using System;
 
-namespace FabricHealer
+namespace FabricHealer.Attributes
 {
     [AttributeUsage(AttributeTargets.Assembly, AllowMultiple = true)]
-    public class CustomServiceInitializerAttribute: Attribute
+    public class CustomServiceInitializerAttribute : Attribute
     {
         public Type InitializerType { get; }
 
         public CustomServiceInitializerAttribute(Type startupType)
         {
-            this.InitializerType = startupType;
+            InitializerType = startupType;
         }
     }
 }
